repos/offchain_tx_rollup/requirements.txt

web3==5.31.1
eth-account==0.5.9
fastapi==0.68.0
uvicorn==0.15.0
pytest==6.2.5
python-dotenv==0.19.0
